Fixed parameter tests in parameters.scala
 Please enter the commit message for your changes. Lines starting
